refactor(studio): replace ChatBar sidebar with ChatPage routing
Route book/book-create pages to the new ChatPage component instead of
BookDetail/BookCreate. Remove the ChatPanel sidebar toggle and delete
ChatBar.tsx. Extract its pure utility functions to chat-utils.ts so the
existing tests continue to pass.
